script/requirements.txt
paramiko>=3.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiosqlite>=0.19.0 